Remove wati webhook handler
