+++ pkg/auth/handler/webapp/deps.go
@@ -75,8 +75,6 @@ var DependencySet = wire.NewSet(
 	wire.Struct(new(WebsocketHandler), "*"),
 	wire.Struct(new(WechatAuthHandler), "*"),
 	wire.Struct(new(WechatCallbackHandler), "*"),
-	wire.Struct(new(WhatsappWATICallbackHandler), "*"),
-	NewWhatsappWATICallbackHandlerLogger,
 	wire.Struct(new(PasskeyCreationOptionsHandler), "*"),
 	wire.Struct(new(PasskeyRequestOptionsHandler), "*"),
 	wire.Struct(new(ConnectWeb3AccountHandler), "*"),
