+++ pkg/auth/wire_handler.go
@@ -294,13 +294,6 @@ func newWebAppWhatsappOTPHandler(p *deps.RequestProvider) http.Handler {
 	))
 }
 
-func newWhatsappWATICallbackHandler(p *deps.RequestProvider) http.Handler {
-	panic(wire.Build(
-		DependencySet,
-		wire.Bind(new(http.Handler), new(*handlerwebapp.WhatsappWATICallbackHandler)),
-	))
-}
-
 func newWebAppSetupLoginLinkOTPHandler(p *deps.RequestProvider) http.Handler {
 	panic(wire.Build(
 		DependencySet,
